update cortana host configs and remove addtoany
