+++ source/nodejs/adaptivecards-designer/index.html
@@ -32,17 +32,8 @@
 			// This is not required. When no list is passed (empty array or null), the designer
 			// uses a default built-in host container, and the host container picker in the
 			// toolbar is hidden.
-			let hostContainers = [];
-
-			hostContainers.push(new ACDesigner.WebChatContainer("Bot Framework WebChat", "containers/webchat-container.css"));
-			hostContainers.push(new ACDesigner.CortanaContainer("Cortana Skills", "containers/cortana-container.css"));
-			hostContainers.push(new ACDesigner.OutlookContainer("Outlook Actionable Messages", "containers/outlook-container.css"));
-			hostContainers.push(new ACDesigner.TimelineContainer("Windows Timeline", "containers/timeline-container.css"));
-			hostContainers.push(new ACDesigner.DarkTeamsContainer("Microsoft Teams - Dark", "containers/teams-container-dark.css"));
-			hostContainers.push(new ACDesigner.LightTeamsContainer("Microsoft Teams - Light", "containers/teams-container-light.css"));
-			hostContainers.push(new ACDesigner.BotFrameworkContainer("Bot Framework Other Channels (Image render)", "containers/bf-image-container.css"));
-			hostContainers.push(new ACDesigner.ToastContainer("Windows Notifications (Preview)", "containers/toast-container.css"));
-
+			let hostContainers = ACDesigner.defaultMicrosoftHosts;
+			
 			let designer = new ACDesigner.CardDesigner(hostContainers);
 
 			// Change where the designer looks for assets like CSS and container images
