+++ source/nodejs/adaptivecards-designer/previewFeatures.html
@@ -27,19 +27,7 @@
 				console.log("Local storage is not available.");
 			}
 
-
-			let hostContainers = [];
-
-			hostContainers.push(new ACDesigner.WebChatContainer("Bot Framework WebChat", "containers/webchat-container.css"));
-			hostContainers.push(new ACDesigner.CortanaContainer("Cortana Skills", "containers/cortana-container.css"));
-			hostContainers.push(new ACDesigner.OutlookContainer("Outlook Actionable Messages", "containers/outlook-container.css"));
-			hostContainers.push(new ACDesigner.TimelineContainer("Windows Timeline", "containers/timeline-container.css"));
-			hostContainers.push(new ACDesigner.DarkTeamsContainer("Microsoft Teams - Dark", "containers/teams-container-dark.css"));
-			hostContainers.push(new ACDesigner.LightTeamsContainer("Microsoft Teams - Light", "containers/teams-container-light.css"));
-			hostContainers.push(new ACDesigner.BotFrameworkContainer("Bot Framework Other Channels (Image render)", "containers/bf-image-container.css"));
-			hostContainers.push(new ACDesigner.ToastContainer("Windows Notifications (Preview)", "containers/toast-container.css"));
-
-			let designer = new ACDesigner.CardDesigner(hostContainers);
+			let designer = new ACDesigner.CardDesigner(ACDesigner.defaultMicrosoftHosts);
 			require.config({
 				paths: {
 					'vs': 'https://unpkg.com/monaco-editor@0.17.1/min/vs'
