+++ source/nodejs/adaptivecards-designer/src/containers/cortana-classic/cortana-bottomleft.png
