+++ source/nodejs/adaptivecards-designer/src/containers/cortana-classic/cortana-frame.png
