+++ source/nodejs/adaptivecards-designer/src/containers/cortana-classic/cortana-magnifyingglass.png
