+++ source/nodejs/adaptivecards-designer/src/containers/index.ts
@@ -3,7 +3,8 @@
 export * from "./host-container";
 export * from "./default/default-container";
 export * from "./bf-image/bf-image-container";
-export * from "./cortana/cortana-container";
+export * from "./cortana-classic/cortana-classic-container";
+export * from "./cortana/cortana-container"
 export * from "./notifications/toast-container";
 export * from "./outlook/outlook-container";
 export * from "./skype/skype-container";
