+++ source/nodejs/adaptivecards-site/themes/adaptivecards/layout/_partial/footer.ejs
@@ -109,7 +109,7 @@
 											our ads</a>
 									</li>
 
-									<li>© Microsoft 2019</li>
+									<li>© Microsoft 2020</li>
 
 								</ul>
 							</nav>
