+++ source/nodejs/adaptivecards-site/themes/adaptivecards/layout/_partial/head.ejs
@@ -20,8 +20,6 @@
 	<title><% if (title){ %><%= title %> | <% } %><%= config.title %></title>
 	<meta name="viewport" content="width=device-width, initial-scale=1">
 	
-
-
 	<link rel="alternate" href="<%- url_for(config.feed.path) %>" title="<%= config.title %>" type="application/atom+xml">
 	<link rel="icon" href="<%- url_for('content/favicon.ico') %>">
