+++ source/nodejs/adaptivecards-site/themes/adaptivecards/layout/post.ejs
@@ -17,7 +17,7 @@
 	}
 </style>
 
-<%- hashed_asset("js/addtoany.js") %>
+
 
 <div class="blog w3-content" style="max-width: 960px">
